+++ Changelog.md
@@ -12,6 +12,7 @@
 - Fix: Placing a Secure Redstone Interface as a receiver does not correctly update itself and surrounding blocks
 - Fix: Blocks projected by a Projector are reset to their default state when opening the Projector's screen
 - Fix: The owner of the topmost block of the cage spawned by a Cage Trap is not set correctly 
+- Fix: Nearby chunks can visually disappear from the world when dismounting a Security Camera that's close to the player
 
 --------------------------Changelog for v1.10 of SecurityCraft--------------------------